Alteração do endpoint para POST
